Add AnalysisTask model.
Add a model that will keep track of all the analysis jobs that are
run on a given dataset.  This is the primary way that a background
job will communicate with the web frontend, it is responsible for
reporting on the status of the job and returning its results to the
user.