Add an export-citations job.
This commit adds the background job for exporting a dataset as
a zip file filled with citations.  For now, we just have one file
per item in the dataset, even for export formats that could be
merged.  This is because it saves us from having to load the
entirety of the result array into memory.