Fix the semantic confusion between job and task.
I had been far too cavalier with conflating 'job' and 'task'
throughout the source base.  Fixing that confusion is important to
maintaining the difference between analysis tasks, which are stored
in the database, and background jobs, which are not persisted at
all.  Job views are correct -- they're tied to particular varieties
of background job.  But there needs to be a connection between the
task instances and the job classes -- hence the 'job_type' parameter
in AnalysisTask.