Citeproc-ruby 0.0.4 works on 1.8.
I worked with the author of citeproc-ruby to patch together a gem that does
well enough under Ruby 1.8 that it passes our tests.  Also, I'm going to
try testing JRuby 1.9 mode on Travis.  It fails locally with a really weird
error message, I'm going to see if I get the same thing remotely.