Add job to export single-doc term vectors.
The term-vector export is particularly easy for a single document,
so we can code that up as its own separate sort of job.  In all
likelihood, we'll be giving you only the top N occurring terms for
large datasets, so this will be the only way to get to the full
TVs.