Add trailing slashes on all URLs, 'data-url'.
jQuery Mobile very much wants us to put a trailing slash on all
"directories" that don't have a filename included.  (See the docs:
http://jquerymobile.com/test/docs/pages/page-navmodel.html.)  At
almost every point on the site, you might get redirected into a
subdirectory, so we're just going to put trailing slashes on
everything and let Rails sort it out, which it will.

Also, add support for the 'data-url' attribute in every page, which
should resolve all the troubles we've been having with redirects.