datasets#add cannot be a member action.
If datasets#add is a member action, then whenever we call
'add_dataset_path' in the search controller, the dataset id is
filled in automatically with whatever params[:id] is currently
set to (the shasum of some document or other).  To fix this, we
just can't get datasets#add be a member action, and we specify
the dataset_id and shasum parameters to that action manually.