Split the word frequency analysis core out to a class.
We want to return the word frequency results as CSV, which is almost
impossible to parse.  So if we're going to unit test the word
frequency analysis code, we've got to break it out into its own class.
This commit does that.  Also, there's still a brief test for the
analysis job.

More still remains to be done for writing a thorough test for the
WordFrequencyAnalyzer.